Introduces an sqlite interface
Uses sqlitedict to implement a partd interface that is backed by sqlite.
The desired goal is to support kv stores that are larger than memory,
run on distributed file systems that will break under the File
interface. (The file interface writes one file per key, the sqlitedict
interface writes a single database file).

Includes basic tests to ensure that the SqliteInterface is useable and
compo sable.